In FullscreenTextButtonPanel, ePixmap should be widget
Change the contents of FullscreenTextButtonPanel from ePixmap to
widget. The contents were incorrectly changed from widget to ePixmap
when the button was moved into a screen template. It needs to be a
widget so that whether it appears or not is under program control.
See, for example, in MENU>Setup>Network>Mount Manager. The TEXT button
should only appear when text entry fields are highlighted.

Fix an incorrect comment in skin_templates.xml.