Some reordering of live TV infobars & MoviePlayer infobar
Moved InfoBar & SecondInfoBar to the head of the file.  Within
InfoBarCommon, moved the bottom infobar background to the beginning
of the bottom infobar, rearranged the tuner information into a more
regular order, and added comments to the sections of the tuner
information. Reordered AGC and SNR information into a consistent
order.

In MoviePlayer, moved the bottom bar background to the start of the
bottom bar layout, some reordering to get left-to-right order in
screen displays.