Adjusted some incorrect dimensions in the InfoBarCommon background.
Reworked background for MoviePlayer infobar to match the layout of
the InfoBarCommon. Some of the dimesnions were out, and it was
missing the top PNG fate from black.

Restored the '-' in the time range in the Now information in
SecondInfoBar, and tweaked the locations so that Now and Next time
range, time remaining & duration have the same field spacing.

Tweaked the spacing of the time range for Now and Next in the bottom
inforbar is more even.

Made the right-hand edge of the SecondInfoBar Now and Next event
name text boxes line up with the right-hand edge of the extended
information box.  This shortened the Now text box (it used to go
to the edge of the panel) and lengthened the Next text box.