Factored out some common elements in Setup screens
SetupCommon for basic screen layout, coloured buttons and help text frame,
SetupTextHelp for the remote text input help box in its usual bottom
left position.
SetupConfig for the majority of setup configuration lists.

It's not practical to place the help text widget into a screen with its frame
without modifications to the Python code for the widgets.

Putting the help text frame in a standard position required the help
text for the Location Scan, Software Manager Setup and General Setup
to be move to be in the same location as for the other help text items.

Changed HarddiskSetup screen's explicit ePixmaps for its coloured button hints into
the coresponding panels. This screen doesn't seem to be used in the current firmware.