Tidying up of some text box sizes and list sizes
Adjust the size of the text area to accommodate more text. This
leaves only two setup screen items that don't fit in the text area,
both in MENU>Setup>TV>Recording Settings:
    Preferred tuner for recordings
    Default recording type

Use SetupCommon & SetupConfig in some screens missed in the first
pass.

Fix some text frames where the size sattribute didn't reflect the
actual frame size, and in some cases adjust their text boxes to fit
the frame size.

Moved the menu in Download/Remove plugins to the same location as
other setup lists, and lengthened it.

Adjust the zPosition of the TEXT button hint in FullscreenTextButtonPanel
so that it reliably overlays the MENU button hint when both are
displayed.

Moved HelpWindow into SetupTextHelp and removed explicit references
to it from setup screens.

Adjust the height of the FileCommander list boxes so that they show
only complete lines.

Added newlines to the end of files that didn't have them.