Tidy up LocationBox and derivitives
In screens LocationBox, MovieLocationBox and TimeshiftLocationBox,
adjust the vertical position and size of widgets filelist, booklist
and textbook.  The adjustments were made so that filelist is an
integral number of items high and 10 items high rather than ~6.5
items. Textbook was moved so it doesn't overlap the bottom of
filelist, and booklist was moved to accommodate the other changes.

The item height in filelist is 23 (set in the code) and the item
height in booklist is 25 (and not modifiable in the code, as far
as I could see).