Fix bug 175 - Grapical EPG, Change timer, Delete hot key is green should be red.
**Prerequisite for Fix-bug-175 in easy-ui-4**

Added screen RecordTimerQuestion1 to allow for two different lengths
(2 & 3) of ChoiceBox lists for GREEN Change Timer and REC in
EPGSelection.ChoiceBox().

Needed because the Change Timer list fix in easy-ui-4 now shows all
three entries for the Change Timer list: Delete timer, Edit timer
and Disable timer, and autosizing the ChoiceBox (which fixes the
vertical dimension) gives a widtn of 520.