Use skin templates in Information screen ServiceInfo and Live TV popup ServiceInfoSimple
In skin_setup.xml:

Use templates in Information screen ServiceInfo.

A new ServiceInfoWidget has been created with its own template,
because it doesn't share any template items with other Information
screens.

Create new widget panel screens for ServiceInfo and Live TV
ServiceInfoSimple suitable for placing as panels in the two screens.

In skin.xml:

Use a panel instead of a widget in ServiceInfoSimple to the
same template as ServiceInfo.

Adjust the panel size and location so that the text isn't hard up
against the left and bottom edges.