[GraphicalEPGs] - Tidy skin in preparation for adding timeline tick marks - tweal skin layouts
[GraphicalEPG] - Add heading bar with bouquet name and date/time
to reflect the similar bar above the timeline in GraphicalEPGPIG.

[GraphicalEPG] - Change the timeline_text colours to be the same
as those used in GraphicalEPGPIG.

[GraphicalEPG] - Move the top bar of the timeline_now cursor up one
pixel so that there's a small gap between it and the top of the
main display area, as in GraphicalEPGPIG.

[GraphicalInfoBarEPG] - Move bottom of itmeline_text up so that it
doesn't overlap the list box.

[GraphicalInfoBarEPG] - Slim down the top bar of the timeline_now
cursor from 4 pixels to 2 pixels to make more room for the timeline
tickmarks.