Fix bug #301 - Help menu in main screen errors - improvements
Change the screen title from "Menu" to "Help".

Added additional widgets to allow the display of up to four active
buttons on the remote image. The widgets also use the "offset" attribute
so that the pixmap can be located relative to its indicator point
(centre of the circle for the current pixmaps).

Allow two lines for display of extra long press and Shift information
when there are multiple buttons displayed on the image.

Add new yellow_circle23x23.png to more cleanly indicate when multiple
keys are active.

Requires corresponding commit in Fix-bug-301 in easy-ui-4 repository.