Fix Bug #301 - Allow for more button indicators in HELP screens
Increase the number of indicators to 16, mainly to allow for
numeric key help where all the buttons have the same description
(e.g. in zap channel by number). This already happens in HELP
for live TV, and without this fix, only 4 of the 9 active buttons
(1-9) are indicated.

Supports corresponding commit in branch Fix-bug-301-10 in
easu-ui-4 repository.