Fix Bug #359 - No "current position" indicator in Cutlist editor timeline
Remove easy-skin-aus-hd from pointer attribute of widget element
Timeline in CutListEditor and (currently unused?) TitleCutter
screens, so that the PNG lookup first searches easy-skin-aus-hd,
then skin_default. This allows it to find the PNG in skin_default,
but it can be overridden by a different one in easy-skin-aus-hd if
necessary.