Fix #368 - Rearrange infobars for more consistent appearance and to save space
Move the main infobar clock into the main infobar panel and lower
the picon to accommodate it. Move the service number to be smaller
text matching the service name, and positioned to the left of the
service name.

Reduce the height of the Movie Player infobar, by moving the replay
state from above the progress bar to the left end of the progress
bar and reduce excessive padding. Height reduced by ~30%.

Move the Movie Player clock and video status info to correspond to
their locations in the main infobar (swap sides).

In both infobars, change the screen resolution display to include
the progressive/interlaced flag and the field rate.