Fix #368 - Allow user control over display of service numbers in infobar
Change source/converter/renderer chains for service name and service
number to widgets to allow their visibility to be controlled by
MENU>Setup>TV>Channel settings>Show channel numbers in channel selection and infobar.
Affects infobar and the "now" panel of the second infobar.

When the service number is hidden in the infobar, a different service
name widget is shown so that the service name "moves" across to use
the space vacated by the service number.

Tidy up alignment in the second infobar "now" and "next" panels,
and fix alignment error in the start and end times in both panels.
The error was probably introduced when the panels were relocated
in the previous commit.

Supports (and requires) corresponding changes made in
Implement-enhancement-368 branch in easy-ui-4 repository.