Fix #391: Show number of items in file list
Show the number of files/recordings and the number of directories
(excluding '..' and 'Trash') in the selection list in the space
above the list.

Add the volume size to the volume usage information below the
selection list.

Move the size/duration information for the recording/media file in
focus to be above the file's details in the right-hand panel.

Align all these information fields to the left.

Centre the "behaviour at end" icon over the TEXT button hint. The
TEXT button defaults to changing the "behaviour at end" setting.
(Small tweak.)

Requires corresponding changes in branch Implement-enhancement-391
in repository easy-ui-4.