Fix bug #414: [MovieSelectionSlim] "Slim" Movie Player screen a bit of a mess
Moved service name & duration of current focus item so it doesn't
overlap with button hints. Mover the "space available" display so
that it doesn't overlap the selection list.

Moved the sort order and "after end of play" icons from off-screen
to on-screen.

Used the sort order and "after end of play" icons from IanSav's
OverlayHD skin instead of the skin default icons, mainly because
the default skin icons don't cover all the Beyonwiz sort options.

The extra icons necessitated putting a new MovieSelectionSlim into
the easy-skin-aus-hd skin.

Moved the synopsis of the current item to on-screen and reduced its
height,

Reduced the height of the selection list to accommodate the other
changes in the same screen height.

The MovieSelectionSlim screen selection list is now a bit tight if
the default 20 rows are displayed. It works well if the number of
rows is reduced to 15.

Corresponding changes (apart from the icons) to the default skin
MovieSelectionSlim have been merged into the oe-alliance-e2-skindefault
repository (commit 550130a0ab).

I have also added the sort order icon to [MovieSelection].