Protect against division by zero
There were reported crashes for a configuration where the start and end
times of a timer are set to the same value. Now we perform a sanity check
on the duration and only use it if the value is greater than zero.

Also simplified three lines of code that only multiplied a value by 100.