Fix bug #532: [skin] Fill anomalies in MoviePlayer Infobar
Match transparency of "abackground" to the transparency of the
navy_fondo-fs8.png fill pixmap.

The fill colour "abackground" is also used in the "now" and "next"
panels of the second infobar screen, but the change in transparency
does not noticeably affect their presentation.

Change height of MoviePlayer infobar screen to match the actual
height of the infobar.

Remove the alphatest="on" attribute from the navy_fondo-fs8.png
pixmap in the MoviePlayer infobar screen to match the pixmap
attributes in InfoBarCommon.