Fix bug #611: [skins] Non-standard page layout and wasted space in IMDb search screens
Rearrange the fields in the IMDB screen skin to use the standard
skin layout and to use all the available space in the screen.

Adjust the poster thumbnail size to match the height of the text
panel beside it and to have the same aspect ratio as the IMDb poster
images and the plugin's no_poster.jpg image.