Fix bug #612: [skins] Improve event information panel in EPGSelection and EPGSearch screen skins
Change description from ExtendedDescription to FullDescription to
get both short and long descriptions if present and different.

Add start/end times, duration, genre and classification to the event
information.

Make the "number" widget conditional to avoid a warning in the
EPG_TYPE_SINGLE EPG type, which doesn't define a "number" widget.

Move and resize the "number" widget to cover the new size of the
event information.

Put the event information into a panel for possible re-use.