Fix issue #613: [skins] Use classification icons instead of abbreviation
Rearranged some layouts slightly so that there is now room for the
genre, rating icon and full rating name. The only screen skin
remainig that still uses the short rating name is GraphicalEPGPIG.

Cut back the length of the description in MovieSelection so that
it no longer overlaps the genre and ratings items.

Thanks to IanSav for permission to use the classification icons
from the OverlayHD skin.