Fix Bug #720: [NetworkBrowser skin] Cleanups required - more adjustments
Adjust all templated lists in the NetworkBrowser to have an item
height of 60 (previously mixed 60 & 70).

Tweak template positions & sizes to fit the new height and eliminate
under-height text boxes and overlaps between boxes.

Adjust all templated lists to have height 480 to allow for 8 items
instead of the previous 7 for the old 420-high lists with item
height 60 and 6 for the same height & item height 70.

Adjust all templated lists to have a top of 120.

Adjust headings in AutoMountView to fit in the reduced gap and put
back the transparent="1" in legend3 that was inadvertently deleted
in the previous commit.